// backend/src/common/enums/operation.code.enum.ts
export enum RoleEnum {
  SUPERADMIN = 'superadmin',
  ADMIN = 'admin',
  USER = 'user',
}

export enum OperationCodeEnum {
  ADD_OPERATION = 'ADD_OPERATION',
  REMOVE_OPERATION = 'REMOVE_OPERATION',
  /// USER SECTION
  GET_USERS = 'GET_USERS',
  UPDATE_USERS_ROLE = 'UPDATE_USERS_ROLE',
  UPDATE_USER_ROLE = 'UPDATE_USER_ROLE',
  UPDATE_NOTIFICATION_HOURS = 'UPDATE_NOTIFICATION_HOURS',
  UPDATE_USER_NICKNAME = 'UPDATE_USER_NICKNAME',

  /// ROLE SECTION
  GET_ROLES = 'GET_ROLES',
  CREATE_ROLE = 'CREATE_ROLE',
  DELETE_ROLE = 'DELETE_ROLE',
  GET_ROLE_OPERATIONS = 'GET_ROLE_OPERATIONS',
  ADD_OPERATION_TO_ROLE = 'ADD_OPERATION_TO_ROLE',
  EDIT_OPERATION_OF_ROLE = 'EDIT_OPERATION_OF_ROLE',
  DELETE_OPERATION_FROM_ROLE = 'DELETE_OPERATION_FROM_ROLE',

  /// OPERATION SECTION
  GET_OPERATIONS = 'GET_OPERATIONS',
  SYNC_OPERATIONS = 'SYNC_OPERATIONS',
  CREATE_OPERATION = 'CREATE_OPERATION',
  EDIT_OPERATION = 'EDIT_OPERATION',
  DELETE_OPERATION = 'DELETE_OPERATION',

  /// MANIFESTATION SECTION
  CREATE_MANIFESTATION = 'CREATE_MANIFESTATION',
  GET_ALL_MANIFESTATIONS = 'GET_ALL_MANIFESTATIONS',
  GET_MANIFESTATION = 'GET_MANIFESTATION',
  UPDATE_MANIFESTATION = 'UPDATE_MANIFESTATION',
  UPDATE_369_MANIFESTATION = 'UPDATE_369_MANIFESTATION',
  DELETE_MANIFESTATION = 'DELETE_MANIFESTATION',
  GET_USER_MANIFESTATIONS = 'GET_USER_MANIFESTATIONS',
}